+++ lib/global/utils/app_update.dart
@@ -7,17 +7,19 @@ class AndroidAppUpdate {
   const AndroidAppUpdate._();
 
   static Future<void> get update async {
-    final res = await InAppUpdate.checkForUpdate();
-    res.updateAvailability == UpdateAvailability.updateAvailable
-        ? InAppUpdate.performImmediateUpdate().then((value) {
-            value == AppUpdateResult.userDeniedUpdate
-                ? SystemNavigator.pop()
-                : null;
-          }).catchError((e) async {
-            Message("Error", e.toString());
-            await Future.delayed(
-                const Duration(seconds: 2), () => SystemNavigator.pop());
-          })
-        : null;
+    try {
+      final res = await InAppUpdate.checkForUpdate();
+      res.updateAvailability == UpdateAvailability.updateAvailable
+          ? InAppUpdate.performImmediateUpdate().then((value) {
+              value == AppUpdateResult.userDeniedUpdate
+                  ? SystemNavigator.pop()
+                  : null;
+            }).catchError((e) async {
+              Message("Error", e.toString());
+              await Future.delayed(
+                  const Duration(seconds: 2), () => SystemNavigator.pop());
+            })
+          : null;
+    } on Exception catch (_) {}
   }
 }